Display and delete comments

+++ csc394/tasks/views.py
@@ -1,17 +1,24 @@
 from django.shortcuts import render
 from django.http import HttpResponse, HttpResponseRedirect
 from .models import TodoItem, Comment
+from django.contrib.auth import authenticate
 
 # Create your views here.
 
 
 def taskView(request):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     all_todo_items = TodoItem.objects.all()
     return render(request, 'tasks.html',
                   {'all_items': all_todo_items})
 
 
 def addTodo(request):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     new_item = TodoItem(
         content=request.POST['content'], title=request.POST['title'],
         date=request.POST['date'], user=request.POST['user'], status='inProgress')
@@ -21,6 +28,9 @@ def addTodo(request):
 
 
 def editTodo(request, todo_id):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     item_to_edit = TodoItem.objects.get(id=todo_id)
     item_to_edit.content = request.POST['content']
     item_to_edit.title = request.POST['title']
@@ -31,6 +41,9 @@ def editTodo(request, todo_id):
 
 
 def deleteTodo(request, todo_id):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     item_to_delete = TodoItem.objects.get(id=todo_id)
     item_to_delete.delete()
     return HttpResponseRedirect('/tasks/')
@@ -38,15 +51,31 @@ def deleteTodo(request, todo_id):
 
 
 def todoDetails(request, todo_id):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     try:
         todoItem = TodoItem.objects.get(pk=todo_id)
     except TodoItem.DoesNotExist:
         return HttpResponse("Task does not exist", status=404)
-    return render(request, 'tasks/taskDetails.html', {'todoItem': todoItem})
+    comments = Comment.objects.filter(todoItem=todo_id)
+    return render(request, 'tasks/taskDetails.html', {'todoItem': todoItem, 'comments': comments})
 
 def addComment(request, todo_id):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
     new_item = Comment(
-        todoItem=TodoItem.objects.get(id = request.POST['todoItem']), title=request.POST['title'], 
-        body=request.POST['body'], user=request.POST['user'])
+        todoItem=TodoItem.objects.get(id = request.POST['todoItem']), 
+        title=request.POST['title'], body=request.POST['body'], 
+        user=request.POST['user'])
     new_item.save()
+    return HttpResponseRedirect('/tasks/{0}'.format(todo_id))
+
+def deleteComment(request, todo_id, comment_id):
+    if not request.user.is_authenticated:
+        return render(request, 'login/signin.html')
+
+    item_to_delete = Comment.objects.get(pk=comment_id)
+    item_to_delete.delete()
     return HttpResponseRedirect('/tasks/{0}'.format(todo_id))
